Upgrade pkginfo library to v1.12.1.2
This is the latest version as of 2025-05-04. The prior version, v1.10.0
cannot understand more modern metadata formats, failing silently without
populating dependencies; which leads to confusing errors during early
runtime when third-party packages fail to resolve imports of their
dependencies. Notably, wheel builds for the latest `pyspark`, v3.5.5,
result in a metadata version of v2.4 which is too new for the older
`pkginfo` library to comprehend.